+++ src/directives/rz-error.ts
@@ -1,5 +1,5 @@
-import type { InsertOperation } from '../core/constants';
-import { getDirectiveValue, resolveInsertOperations } from '../core/shared';
+import type { SwapOperation } from '../core/constants';
+import { getDirectiveValue, resolveSwapOperations } from '../core/shared';
 import type { ConfigDirective, DirectiveSlug } from '../types';
 
 const SLUG = 'error' as const satisfies DirectiveSlug;
@@ -9,8 +9,8 @@ export const rzError = {
   getConfig: (el: Element, appRoot: Element, overrideValue?: string | null) => {
     const value = overrideValue || getDirectiveValue(el, SLUG);
     // Errors must have an explicit target to mutate the DOM, so we return `[]` here
-    // to avoid the default-insert behavior of `resolveInsertOperations`.
+    // to avoid the default swap behavior of `resolveSwapOperations`.
     if (!value?.trim()) return [];
-    return resolveInsertOperations(value, el, appRoot);
+    return resolveSwapOperations(value, el, appRoot);
   },
-} as const satisfies ConfigDirective<InsertOperation[]>;
+} as const satisfies ConfigDirective<SwapOperation[]>;