+++ src/index.ts
@@ -2,7 +2,7 @@ export { type RouseConfig, rouse } from './core/app';
 export { debounce, throttle } from './core/timing';
 export { dispatch, on } from './dom/scheduler';
 export { defineScope } from './dom/scope';
-export { insert } from './dom/utils';
+export { swap } from './dom/swapper';
 export {
   computed,
   effect,